Generate metabolites-exposures network

Generate metabolites-exposures network

getemnet(meta, expo, cutoffm = 0.9, cutoffe = 0.6, nacf = 0.2, ...)

Arguments

meta
metabolites dataframe with row as metabolites/peaks and column as samples

expo
exposures dataframe with row as exposures/peaks and column as samples

cutoffm
correlation coefficient cutoff for metabolites-metabolites network

cutoffe
correlation coefficient cutoff for exposures-exposures network

nacf
NA cutoff for exposure, exposure with NA percentage larger than cutoff will be removed and others will be imputed by knn methods, default 0.2

...
other parameters for `cor`

Value

list with table of metabolites-exposures, metabolites-exposures network clusters data and edge table

Examples

data(expo)
data(meta)
me <- getemnet(exp(meta),expo)

#> 9 metabolites correlation network clusters found

#> 155 out of 177 metabolites have no correlation with other metabolites

#> 169 exposures could be used for network analysis

#> 15 exposures correlation network clusters found

#> 91 out of 169 exposures have no correlation with other exposures

#> 170 peaks with multiple exposures associated

#> 1 peaks with 83 exposures associated

#> 4 peaks with 79 exposures associated

#> 4 peaks with 35 exposures associated

#> 2 peaks with 45 exposures associated

#> 3 peaks with 52 exposures associated

#> 6 peaks with 50 exposures associated

#> 3 peaks with 103 exposures associated

#> 1 peaks with 14 exposures associated

#> 4 peaks with 53 exposures associated

#> 1 peaks with 76 exposures associated

#> 3 peaks with 70 exposures associated

#> 3 peaks with 66 exposures associated

#> 2 peaks with 2 exposures associated

#> 2 peaks with 49 exposures associated

#> 2 peaks with 26 exposures associated

#> 7 peaks with 1 exposures associated

#> 2 peaks with 88 exposures associated

#> 6 peaks with 40 exposures associated

#> 3 peaks with 23 exposures associated

#> 2 peaks with 24 exposures associated

#> 5 peaks with 12 exposures associated

#> 2 peaks with 31 exposures associated

#> 2 peaks with 90 exposures associated

#> 2 peaks with 63 exposures associated

#> 2 peaks with 34 exposures associated

#> 1 peaks with 21 exposures associated

#> 2 peaks with 11 exposures associated

#> 2 peaks with 20 exposures associated

#> 4 peaks with 68 exposures associated

#> 2 peaks with 67 exposures associated

#> 1 peaks with 120 exposures associated

#> 4 peaks with 39 exposures associated

#> 1 peaks with 6 exposures associated

#> 3 peaks with 58 exposures associated

#> 3 peaks with 54 exposures associated

#> 1 peaks with 32 exposures associated

#> 3 peaks with 48 exposures associated

#> 3 peaks with 85 exposures associated

#> 2 peaks with 30 exposures associated

#> 1 peaks with 77 exposures associated

#> 2 peaks with 64 exposures associated

#> 1 peaks with 81 exposures associated

#> 3 peaks with 3 exposures associated

#> 1 peaks with 100 exposures associated

#> 2 peaks with 57 exposures associated

#> 1 peaks with 80 exposures associated

#> 4 peaks with 73 exposures associated

#> 2 peaks with 15 exposures associated

#> 2 peaks with 82 exposures associated

#> 3 peaks with 59 exposures associated

#> 1 peaks with 41 exposures associated

#> 4 peaks with 91 exposures associated

#> 2 peaks with 44 exposures associated

#> 2 peaks with 124 exposures associated

#> 1 peaks with 126 exposures associated

#> 1 peaks with 112 exposures associated

#> 2 peaks with 109 exposures associated

#> 1 peaks with 125 exposures associated

#> 1 peaks with 96 exposures associated

#> 2 peaks with 9 exposures associated

#> 1 peaks with 121 exposures associated

#> 2 peaks with 102 exposures associated

#> 2 peaks with 56 exposures associated

#> 1 peaks with 46 exposures associated

#> 1 peaks with 25 exposures associated

#> 1 peaks with 84 exposures associated

#> 1 peaks with 97 exposures associated

#> 1 peaks with 65 exposures associated

#> 2 peaks with 51 exposures associated

#> 1 peaks with 127 exposures associated

#> 1 peaks with 60 exposures associated

#> 1 peaks with 16 exposures associated

#> 2 peaks with 36 exposures associated

#> 3 peaks with 33 exposures associated

#> 1 peaks with 110 exposures associated

#> 1 peaks with 98 exposures associated

#> 1 peaks with 19 exposures associated

#> 1 peaks with 7 exposures associated

#> 1 peaks with 29 exposures associated

#> 1 peaks with 106 exposures associated

#> 1 peaks with 22 exposures associated

#> 1 peaks with 8 exposures associated

#> 1 peaks with 17 exposures associated

#> 1 peaks with 55 exposures associated

#> 1 peaks with 72 exposures associated

#> 1 peaks with 61 exposures associated

#> 22 peaks connected with other peaks and exposures

#> 155 peaks connected with exposures only