Generate metabolites-exposures network

Generate metabolites-exposures network

getemnet(meta, expo, cutoffm = 0.9, cutoffe = 0.6, nacf = 0.2, ...)

Arguments

meta

metabolites dataframe with row as metabolites/peaks and column as samples

expo

exposures dataframe with row as exposures/peaks and column as samples

cutoffm

correlation coefficient cutoff for metabolites-metabolites network

cutoffe

correlation coefficient cutoff for exposures-exposures network

nacf

NA cutoff for exposure, exposure with NA percentage larger than cutoff will be removed and others will be imputed by knn methods, default 0.2

...

other parameters for `cor`

Value

list with table of metabolites-exposures, metabolites-exposures network clusters data and edge table

Examples

data(expo)
data(meta)
me <- getemnet(exp(meta),expo)
#> 9 metabolites correlation network clusters found
#> 155 out of 177 metabolites have no correlation with other metabolites
#> 7 exposures could be used for network analysis
#> exposures have no correlation with each others.
#> 116 peaks with multiple exposures associated
#> 14 peaks with 5 exposures associated
#> 27 peaks with 3 exposures associated
#> 32 peaks with 1 exposures associated
#> 49 peaks with 2 exposures associated
#> 13 peaks with 4 exposures associated
#> 13 peaks with 6 exposures associated
#> 14 peaks connected with other peaks and exposures
#> 134 peaks connected with exposures only